fix: address code review feedback - architecture, error handling, SQL escaping
Agent-Logs-Url: https://github.com/martin-macak/guru/sessions/1d861f4e-752c-4a3a-ba2e-e00e8474eb64
