+++ packages/guru-core/src/guru_core/__init__.py
@@ -14,12 +14,19 @@ from guru_core.types import (
     StatusOut,
     StatusResponse,
 )
+from guru_core.config import (
+    DEFAULT_RULES,
+    load_rules,
+    merge_rules,
+    resolve_config,
+)
 from guru_core.discovery import find_guru_root, GuruNotFoundError
 from guru_core.client import GuruClient
 from guru_core.autostart import ensure_server, ServerStartError
 
 __all__ = [
     "ChunkingConfig",
+    "DEFAULT_RULES",
     "DocumentInfo",
     "DocumentListItem",
     "DocumentOut",
@@ -38,4 +45,7 @@ __all__ = [
     "StatusResponse",
     "ensure_server",
     "find_guru_root",
+    "load_rules",
+    "merge_rules",
+    "resolve_config",
 ]
